Refractor index and App.js
- Setup Enzyme and first test for app.js
