+++ src/App.js
@@ -1,55 +1,22 @@
 import './App.css';
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import { BrowserRouter as Switch, Route, useRouteMatch } from 'react-router-dom';
+import { BrowserRouter as Switch, Route, Redirect } from 'react-router-dom';
 
-import Header from './components/Header'
-import Confirmation from './pages/Check-in/Confirmation';
-import Checkin from './pages/Check-in/Checkin'
-
-
-const useStyles = makeStyles({
-  root: {
-    height: '100%',
-  },
-  body: {
-    display: 'grid',
-    gridGap: '2em',
-    padding: '2em 20px',
-    backgroundColor: '#fcefee'
-  },
-  footer: {
-    backgroundColor: '#fc5c9c',
-    color: '#fcefee',
-    width: '100%',
-    textAlign: 'center',
-    padding: '10px 0px',
-    bottom: 0,
-  },
-  adminBody: {
-    height: '100%',
-    backgroundColor: '#FDFDFD'
-  }
-})
+import Client from './pages/Check-in/Client';
+import PrivateRoute from './components/PrivateRoute';
+import Admin from './pages/Admin/Admin';
+import Login from './components/Login';
 
 function App() {
-  const classes = useStyles();
-
-  let { path } = useRouteMatch();
-
   return (
-    <>
-      <Header />
-      <div className={classes.body}>
-        <Switch>
-          <Route exact path={path} component={Checkin} />
-          <Route path={`${path}/confirm`} component={Confirmation} />
-        </Switch>
-      </div>
-      <div className={classes.footer}>
-          <span>Copyright © 2021, Polish Nail, Inc. "Polish Nail" and logo are registered trademarks of Polish Nail, Inc</span>
-      </div>
-    </>
+    <div data-test="component-app" style={{height: '100%'}}>
+      <Switch>
+        <Route exact path="/"><Redirect to="/login" /></Route>
+        <Route path='/checkin' component={Client} />
+        <Route path="/login" component={Login} />
+        <PrivateRoute path="/dashboard" component={Admin} />
+      </Switch>
+    </div>
   );
 }
