+++ src/index.js
@@ -3,12 +3,8 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
-import { BrowserRouter, Switch, Route, Redirect } from 'react-router-dom'
+import { BrowserRouter } from 'react-router-dom'
 import { Auth0Provider } from '@auth0/auth0-react';
-
-import Admin from './pages/Admin/Admin';
-import Login from './components/Login';
-import PrivateRoute from './components/PrivateRoute';
 
 ReactDOM.render(
   <Auth0Provider
@@ -17,12 +13,7 @@ ReactDOM.render(
     redirectUri={`${window.location.origin}/dashboard`}
   >
     <BrowserRouter>
-      <Switch>
-        <Route exact path="/"><Redirect to="/login" /></Route>
-        <Route path="/checkin" component={App} />
-        <Route path="/login" component={Login} />
-        <PrivateRoute path="/dashboard" component={Admin} />
-      </Switch>
+      <App />
     </BrowserRouter>
   </Auth0Provider>,
   document.getElementById('root')
