+++ src/tests/App.test.js
@@ -1,8 +1,22 @@
-import { render, screen } from '@testing-library/react';
-import App from '../App';
+import React from 'react'
 
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
+import { mount } from 'enzyme'
+import App from '../App'
+import { findByTestAttr } from '../../test/testUtils'
+import BrowserRouter from 'react-router'
+
+const setup = () => {
+  return mount(<App />)
+}
+describe('it renders', () => {
+  const mockUseRouteMatch = {
+    path: '',
+    location: ''
+  }
+
+  test('renders without crashing', () => {
+    const wrapper = setup()
+    const app = findByTestAttr(wrapper, 'component-app')
+    expect(app.exists()).toBe(true)
+  })
+})
